Finnished edit blog and event errors for static params
