--- src/app/[...not_found]/page.tsx
@@ -1,12 +1,0 @@
-/*import { notFound } from "next/navigation"
-
-export async function generateStaticParams() {
-    // Generate static paths here, if needed
-    return [];
-  }
-
-export default function error404(){
-    return(
-            notFound()
-    )
-}*/
